Experimented with caching images during load and displaying synchronously - led to immediate image render on table load but also a little scroll lag. Was not implemented.
Changed title colours to Imperial Blue. Colour set is dynamic so it is Imperial blue in light mode and white in dark mode.

TODO? - GIFs not all the same size so the preview rect is slightly too large for some of them. This is more noticeable in dark mode. Not a priority.